ai-agent/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv

# --- LangChain Version-Locked Dependencies ---
# This ensures all imports will work correctly.
langchain==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.7